WEB_PROJECT/requirements.txt
alembic==1.7.7
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-JWT-Extended==4.3.1
Flask-JWT-Simple==0.0.3
Flask-Login==0.6.0
Flask-RESTful==0.3.9
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
Pillow==9.1.0
PyJWT==2.3.0
pytz==2022.1
requests==2.27.1
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.35
SQLAlchemy-serializer==1.4.1
urllib3==1.26.9
Werkzeug==2.1.1
WTForms==3.0.1
python-dotenv==0.19.2
psycopg2-binary==2.9.3